requirements.txt
Flask==3.1.1
PyYAML==6.0.1
Flask-SocketIO==5.5.1
python-docx
docxtpl
PyMuPdf
jsonschema
bibtexparser
pybtex
docx2pdf